refactor: split strategies into separate scripts
